Silence a bunch of annoying compiler warnings

+++ software/firmware/libjpeg-turbo-2.0.3/jdphuff.c
@@ -211,7 +211,7 @@ start_pass_phuff_decoder(j_decompress_ptr cinfo)
 
 #define NEG_1  ((unsigned)-1)
 #define HUFF_EXTEND(x, s) \
-  ((x) < (1 << ((s) - 1)) ? (x) + (((NEG_1) << (s)) + 1) : (x))
+  ((x) < (1 << ((s) - 1)) ? (x) + (int)(((NEG_1) << (s)) + 1) : (x))
 
 #else
